1) Added Pokemon Sapphire / Ruby (any region) proper Save Format [Flash 128K]. Will prevent the "Save is Corrupted" screen. You can start a new game and the savefile will be generated correctly. On the other hand if you don't want to lose your savefile, but you have the former error screen, keep reading!
2) Added Save-Fix mechanism: Since the only games I could test was Pokemon Sapphire /Ruby... if you have a savefile from Pokemon Sapphire / Ruby that shows "Save is Corrupted" screen whose filesize is 64K (ie: Corrupted savefile, but still readable), gbaemu4ds will fix that savefile for you into a fully correct CRC save.

Case scenario: make sure to use the same filename for both the gba file and the sav file (say, 64K sav file). Boot up gbaemu4ds and run the gba file. If the save was detected as corrupted, the DS console will show a "SaveFix" message, mentioning, that you must save in-game, and then, press Y to write back to file the fixed savefile.